+++ packages/core/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to the library will be documented in this file.
 
+## vX.X.X (Month DD, YYYY)
+
+- Add React-specific `SubmitHandler` type using `FormEvent` type from React
+- Rename `vanilla` build target in favor of dedicated `react` target
+
 ## v0.4.5 (December 12, 2025)
 
 - Fix `initialInput` passed to `initializeFieldStore` for nullish wrapper schemas (pull request #48)